add comment

+++ python_introduction/rectangle_area.py
@@ -1,6 +1,9 @@
+# declare variable
 length = 10
 width = 5
 
-area = int(length * width)
+#calculate area
+area = length * width
 
+#print area
 print("The area of the rectangle is: ", area)
